+++ tests/external_urllib_request/conftest.py
@@ -8,8 +8,8 @@ from collections.abc import Iterator
 import pytest
 from wrapture import Tape, instrumentation, timeline
 
-from tests.external_urllib.server import Server, serve
-from wrapture_instrumentation.external_urllib import UrllibInstrumentation
+from tests.external_urllib_request.server import Server, serve
+from wrapture_instrumentation.external_urllib_request import UrllibInstrumentation
 
 
 @pytest.fixture