+++ tests/external_urllib_request/test_propagation.py
@@ -10,8 +10,8 @@ import urllib.request
 import wrapture
 from wrapture import Tape, instrumentation, timeline
 
-from tests.external_urllib.server import Server
-from wrapture_instrumentation.external_urllib import UrllibInstrumentation
+from tests.external_urllib_request.server import Server
+from wrapture_instrumentation.external_urllib_request import UrllibInstrumentation
 
 TRACEPARENT = re.compile(r"^00-([0-9a-f]{32})-([0-9a-f]{16})-[0-9a-f]{2}$")