+++ tests/external_urllib_request/test_requests.py
@@ -12,8 +12,8 @@ from typing import Any
 import pytest
 from wrapture import Event, Tape, instrumentation, timeline
 
-from tests.external_urllib.server import Server
-from wrapture_instrumentation.external_urllib import UrllibInstrumentation
+from tests.external_urllib_request.server import Server
+from wrapture_instrumentation.external_urllib_request import UrllibInstrumentation
 
 OPEN = "urllib.request:OpenerDirector.open"